fix: simplify maxent preparation

+++ alftools/maxent.py
@@ -5,7 +5,6 @@
 # Copyright (c) 2022, Dylan Jones
 
 import os
-import re
 import shutil
 import logging
 from .utils import ALF_DIR, call
@@ -13,29 +12,28 @@ from .utils import ALF_DIR, call
 logger = logging.getLogger(__name__)
 
 
-def copy_parameters(src_dir, dst_dir):
+def copy_parameters(src_dir, dst_dir, overwrite=False):
     dst = os.path.join(dst_dir, "parameters")
     src = os.path.join(src_dir, "parameters")
     if os.path.exists(dst):
-        os.remove(dst)
+        if overwrite:
+            os.remove(dst)
+        else:
+            return
     logger.debug("Copying parameters to %s", dst_dir)
     shutil.copy(src, dst)
 
 
-def prepare_green_maxent(root):
-    # Copy parameter file to directories
-    regex = re.compile(r"Green_\d+\.\d{2}_\d+\.\d{2}")
-    dirs = list()
-    logger.info("Preparing directory %s for MaxEnt", root)
-    for name in os.listdir(root):
-        if regex.match(name) is not None:
-            green_dir = os.path.join(root, name)
-            copy_parameters(root, green_dir)
-            dirs.append(green_dir)
-    return dirs
+def run_maxent(directory, name, verbose=False):
+    path = os.path.join(directory, name)
+    # Prepare directory
+    src = os.path.join(directory, "parameters")
+    dst = os.path.join(path, "parameters")
+    if not os.path.exists(dst):
+        logger.debug("Copying parameters to %s", path)
+        shutil.copy(src, dst)
 
-
-def run_maxent(directory, verbose=False):
-    logger.info("Running Max_SAC in %s", directory)
+    # Run MaxEnt
+    logger.info("Running Max_SAC in %s", path)
     cmd = os.path.join(ALF_DIR, "Analysis", "Max_SAC.out")
-    call(cmd, directory, verbose)
+    call(cmd, path, verbose)
